Merge pull request #57 from etalab/schema-generation

+++ csv_detective/schema_generation.py
@@ -3,8 +3,95 @@ from datetime import datetime
 import json
 import os
 import tempfile
+from typing import Optional
 
 from csv_detective.s3_utils import get_s3_client, download_from_minio, upload_to_minio
+
+
+def get_description(format: str) -> str:
+    """Returns generic description for specific field"""
+    format_to_desc = {
+        "adresse": "Adresse",
+        "code_commune_insee": "Le code INSEE de la commune",
+        "code_departement": "Le code INSEE du département",
+        "code_region": "Le code INSEE de la région",
+        "code_fantoir": "Le code FANTOIR de la voie ou du lieu-dit",
+        "code_postal": "Le code postal",
+        "commune": "Le nom de la commune",
+        "departement": "Le nom du département",
+        "insee_canton": "Le nom du canton",
+        "latitude_l93": "La latitude au format Lambert 93",
+        "latitude_wgs_fr_metropole": (
+            "La latitude au format WGS. Ne concerne que des latitudes "
+            "de la métropole française"
+        ),
+        "longitude_l93": "La longitude au format Lambert 93",
+        "longitude_wgs_fr_metropole": (
+            "La longitude au format WGS. Ne concerne que des longitudes "
+            "de la métropole française"
+        ),
+        "pays": "Le nom du pays",
+        "region": "Le nom de la région",
+        "code_csp_insee": "Le code de Catégorie Socio-professionnel INSEE",
+        "code_rna": "Le code RNA de l'association",
+        "code_waldec": "Le code WALDEC de l'association",
+        "csp_insee": "La catégorie socio-professionnel INSEE",
+        "date_fr": "Data au format français",
+        "sexe": "Le sexe",
+        "siren": "Le numéro SIREN à 9 chiffres de l'entreprise (unité légale)",
+        "siret": "Le numéro SIRET à 14 chiffres de l'établissement d'une entreprise",
+        "tel_fr": "Le numéro de téléphone français",
+        "uai": "Le numéro UAI (Unité Administrative Immatriculée) de l'établissement scolaire",
+        "jour_de_la_semaine": "Le jour de la semaine",
+        "mois_de_annee": "Le mois de l'année",
+        "latitude_wgs": "La latitude au format WGS",
+        "longitude_wgs": "La longitude au format WGS",
+        "latlon_wgs": "Les coordonnées XY (latitude et longitude)",
+        "booleen": "Booléen",
+        "email": "L'adresse couriel (email)",
+        "float": "Nombre flottant (à virgule)",
+        "int": "Nombre entier",
+        "json": "Chaîne de caractère json",
+        "mongo_object_id": "Identifiant de base de donnée Mongo",
+        "twitter": "Compte Twitter",
+        "url": "Adresse URL",
+        "uuid": "Identifiant unique au format UUID",
+        "date": "Date",
+        "datetime_iso": "Date au format datetime (ISO)",
+        "datetime_rfc822": "Date au format datetime (RFC822)",
+        "year": "Année",
+    }
+    return format_to_desc.get(format, "")
+
+
+def get_pattern(format: str) -> str:
+    """Returns the pattern for a particular format"""
+    format_to_pattern = {
+        "siren": r"^\d{9}$",
+        "siret": r"^\d{14}$",
+        "code_commune_insee": r"^([013-9]\d|2[AB1-9])\d{3}$",
+        "code_postal": r"^([013-9]\d|2[AB1-9])\d{3}$",
+        "code_departement": r"^(([013-9]\d|2[AB1-9])$|9\d{2}$)",
+        "code_region": r"^\d{2}$",
+        "code_rna": r"^[wW]\d{9}$",
+        "code_waldec": (
+            r"^\d{3}\D\d{1,10}$|^\d\D\d\D\d{10}$|^\d{3}\D{3}\d{1,10}$|^\d{3}\D\d{4}\D\d{1,10}"
+            r"$|^\d{3}\D\d{2}[-]\d{3}$|^\d\D\d\D\d{2}\D\d{1,8}$"
+        ),
+        "uai": r"^(0[0-8][0-9]|09[0-5]|9[78][0-9]|[67]20)[0-9]{4}[A-Z]$",
+        "email": r"^\w+@[a-zA-Z_]+?\.[a-zA-Z]{2,3}$",
+        "twitter": r'^@[A-Za-z0-9_]+$',
+        "mongo_object_id": r'^[0-9a-fA-F]{24}$',
+        "uuid": r'^[{]?[0-9a-fA-F]{8}' + '-?([0-9a-fA-F]{4}-?)' + '{3}[0-9a-fA-F]{12}[}]?$',
+        "url": (
+            r'^https?:\/\/(?:www\.)?[-a-zA-Z0-9@:%._\+~#=]{1,256}\.[a-zA-Z0-9()]'
+            r'{1,6}\b(?:[-a-zA-Z0-9()@:%_\+.~#?&\/=]*)$'
+        )
+    }
+    if format in format_to_pattern:
+        return {"pattern": format_to_pattern[format]}
+    else:
+        return {}
 
 
 def get_validata_type(format: str) -> str:
@@ -15,7 +102,6 @@ def get_validata_type(format: str) -> str:
         "float": "number",
         "string": "string",
         "date": "date",
-        "date_fr": "date",
         "datetime_iso": "datetime",
         "datetime_rfc822": "datetime",
         "json_geojson": "geojson",
@@ -92,17 +178,8 @@ def get_example(format: str) -> str:
 
 def get_constraints(format: str) -> dict:
     """Returns the constraints for a given format"""
+    pattern_constraints = get_pattern(format)
     extra_constraints = {}
-    if format == "code_commune_insee":
-        extra_constraints = {"pattern": "^[0-9]{5}$"}
-    if format == "code_departement":
-        extra_constraints = {
-            "pattern": "^(0[13-9]|[1-8][0-9]|9[0-6]|2[a-bA-B]|97[1-6])$"
-        }
-    if format == "code_postal":
-        extra_constraints = {"pattern": "^[0-9]{5}$"}
-    if format == "code_fantoir":
-        extra_constraints = {"pattern": "^[0-9A-Z][0-9]{3}[ABCDEFGHJKLMNPRSTUVWXYZ]$"}
     if format == "latitude_l93":
         extra_constraints = {"minimum": 6037008, "maximum": 7230728}
     if format == "longitude_l93":
@@ -111,25 +188,24 @@ def get_constraints(format: str) -> dict:
         extra_constraints = {"minimum": 41.3, "maximum": 51.3}
     if format == "longitude_wgs_fr_metropole":
         extra_constraints = {"minimum": -5.5, "maximum": 9.8}
-    if format == "siren":
-        extra_constraints = {"pattern": "^[0-9]{9}$"}
-    if format == "siret":
-        extra_constraints = {"pattern": "^[0-9]{14}$"}
-    return {"required": False, **extra_constraints}
+
+    return {"required": False, **pattern_constraints, **extra_constraints}
 
 
 def generate_table_schema(
     analysis_report: dict,
-    netloc: str,
-    bucket: str,
-    key: str,
-    minio_user: str,
-    minio_pwd: str,
-) -> None:
+    save_file: bool,
+    netloc: Optional[str] = None,
+    bucket: Optional[str] = None,
+    key: Optional[str] = None,
+    minio_user: Optional[str] = None,
+    minio_pwd: Optional[str] = None,
+) -> dict:
     """Generates a table schema from the analysis report
 
     Args:
         analysis_report (dict): The analysis report from csv_detective
+        save_file (bool): indicate if schema should be saved into minio or just returned
         netloc (str): The netloc of the minio instance to upload the tableschema
         bucket (str): The bucket to save the schema in
         key (str): The key to save the schema in (without extension as we will append
@@ -142,61 +218,16 @@ def generate_table_schema(
     fields = [
         {
             "name": header,
-            "description": "",
+            "description": get_description(field_report["format"]),
             "example": get_example(field_report["format"]),
             "type": get_validata_type(field_report["format"]),
-            "constraints": {"required": False},
+            "formatFR": field_report["format"],
+            "constraints": get_constraints(field_report["format"])
         }
         for header, field_report in analysis_report["columns"].items()
     ]
 
-    # Create bucket if does not exist
-    client = get_s3_client(netloc, minio_user, minio_pwd)
-    try:
-        client.head_bucket(Bucket=bucket)
-    except ClientError:
-        client.create_bucket(Bucket=bucket)
-
-    tableschema_objects = client.list_objects(Bucket=bucket, Prefix=key, Delimiter="/")
-    if "Contents" in tableschema_objects:
-        tableschema_keys = [
-            tableschema["Key"]
-            for tableschema in client.list_objects(
-                Bucket=bucket, Prefix=key, Delimiter="/"
-            )["Contents"]
-        ]
-        tableschema_versions = [
-            os.path.splitext(tableschema_key)[0].split("_")[-1]
-            for tableschema_key in tableschema_keys
-        ]
-        latest_version = max(tableschema_versions)
-
-        with tempfile.NamedTemporaryFile() as latest_schema_file:
-            with open(latest_schema_file.name, "w") as fp:
-                download_from_minio(
-                    netloc,
-                    bucket,
-                    f"{key}_{latest_version}.json",
-                    latest_schema_file.name,
-                    minio_user,
-                    minio_pwd,
-                )
-                # Check if files are different
-                with open(latest_schema_file.name, "r") as fp:
-                    latest_schema = json.load(fp)
-                    if latest_schema["fields"] != fields:
-                        latest_version_split = latest_version.split(".")
-                        new_version = (
-                            latest_version_split[0]
-                            + "."
-                            + latest_version_split[1]
-                            + "."
-                            + str(int(latest_version_split[2]) + 1)
-                        )
-                    else:
-                        return None
-    else:
-        new_version = "0.0.1"
+    new_version = "0.0.1"
 
     schema = {
         "$schema": "https://frictionlessdata.io/schemas/table-schema.json",
@@ -205,17 +236,26 @@ def generate_table_schema(
         "description": "",
         "countryCode": "FR",
         "homepage": "",
-        "path": "",
-        "resources": [{"title": "", "path": ""}],
-        "sources": [],
+        "path": "https://github.com/etalab/csv-detective",
+        "resources": [],
+        "sources": [
+            {
+                "title": "Spécification Tableschema",
+                "path": "https://specs.frictionlessdata.io/table-schema"
+            },
+            {
+                "title": "schema.data.gouv.fr",
+                "path": "https://schema.data.gouv.fr"
+            }
+        ],
         "created": datetime.today().strftime("%Y-%m-%d"),
         "lastModified": datetime.today().strftime("%Y-%m-%d"),
         "version": new_version,
         "contributors": [
             {
                 "title": "Table schema bot",
-                "email": "",
-                "organisation": "Etalab",
+                "email": "schema@data.gouv.fr",
+                "organisation": "data.gouv.fr",
                 "role": "author",
             },
         ],
@@ -223,13 +263,71 @@ def generate_table_schema(
         "missingValues": [""],
     }
 
-    tableschema_file = tempfile.NamedTemporaryFile(delete=False)
-    with open(tableschema_file.name, "w") as fp:
-        json.dump(schema, fp, indent=4)
+    if not save_file:
+        return schema
 
-    new_version_key = f"{key}_{new_version}.json"
-    upload_to_minio(
-        netloc, bucket, new_version_key, tableschema_file.name, minio_user, minio_pwd
-    )
-    os.unlink(tableschema_file.name)
-    return {"netloc": netloc, "bucket": bucket, "key": new_version_key}
+    if save_file:
+        if not all([netloc, key, bucket, minio_user, minio_pwd]):
+            raise Exception(
+                "To save schema into minio, parameters : netloc, key, bucket, "
+                "minio_user, minio_pwd should be provided"
+            )
+
+        # Create bucket if does not exist
+        client = get_s3_client(netloc, minio_user, minio_pwd)
+        try:
+            client.head_bucket(Bucket=bucket)
+        except ClientError:
+            client.create_bucket(Bucket=bucket)
+
+        tableschema_objects = client.list_objects(Bucket=bucket, Prefix=key, Delimiter="/")
+        if "Contents" in tableschema_objects:
+            tableschema_keys = [
+                tableschema["Key"]
+                for tableschema in client.list_objects(
+                    Bucket=bucket, Prefix=key, Delimiter="/"
+                )["Contents"]
+            ]
+            tableschema_versions = [
+                os.path.splitext(tableschema_key)[0].split("_")[-1]
+                for tableschema_key in tableschema_keys
+            ]
+            latest_version = max(tableschema_versions)
+
+            with tempfile.NamedTemporaryFile() as latest_schema_file:
+                with open(latest_schema_file.name, "w") as fp:
+                    download_from_minio(
+                        netloc,
+                        bucket,
+                        f"{key}_{latest_version}.json",
+                        latest_schema_file.name,
+                        minio_user,
+                        minio_pwd,
+                    )
+                    # Check if files are different
+                    with open(latest_schema_file.name, "r") as fp:
+                        latest_schema = json.load(fp)
+                        if latest_schema["fields"] != fields:
+                            latest_version_split = latest_version.split(".")
+                            new_version = (
+                                latest_version_split[0]
+                                + "."
+                                + latest_version_split[1]
+                                + "."
+                                + str(int(latest_version_split[2]) + 1)
+                            )
+                        else:
+                            return None
+
+            schema["version"] = new_version
+
+        tableschema_file = tempfile.NamedTemporaryFile(delete=False)
+        with open(tableschema_file.name, "w") as fp:
+            json.dump(schema, fp, indent=4)
+
+        new_version_key = f"{key}_{new_version}.json"
+        upload_to_minio(
+            netloc, bucket, new_version_key, tableschema_file.name, minio_user, minio_pwd
+        )
+        os.unlink(tableschema_file.name)
+        return {"netloc": netloc, "bucket": bucket, "key": new_version_key}
